BLOCKUTILS - noTransform null bug fix BROW - fix for center anchor at prerig not driving handle RIGBLOCKS.factory - no module parent visSub/direct fix

+++ mayaTools/cgm/core/mrs/lib/block_utils.py
@@ -929,7 +929,7 @@ def noTransformNull_verify(self,mode='form',forceNew=False):
         if mNoTransformNull and not forceNew:
             return mNoTransformNull
         
-        if forceNew:
+        if forceNew and mNoTransformNull:
             mNoTransformNull.delete()
         
         str_prerigNull = mc.group(em=True)
